Add a new gitlog2changelog.sh bash script to add a ChangeLog file when using make dist
This bash script's purpose is to convert the 'git log' informations in a
ChangeLog format per version of the tool with respect to the GNU standards
such as informations on authors and date of changes. This script is generic
and could be used for other tools, applications, or others under the
conditions of the GNU General Public License version 2.

For npt case, the ChangeLog file will only be generated when 'make dist'
is called to be added in the npt archive.

Signed-off-by: Raphaël Beamonte <raphael.beamonte@gmail.com>